Move credo AEC specific code to vendor specific code
