+++ sonic_platform_base/sonic_xcvr/xcvr_api_factory.py
@@ -13,9 +13,9 @@ from .api.public.c_cmis import CCmisApi
 from .mem_maps.public.cmis import CmisMemMap
 from .mem_maps.public.c_cmis import CCmisMemMap
 
-from .codes.credo.aec_800g import CmisAec800gCodes
-from .api.credo.aec_800g import CmisAec800gApi
-from .mem_maps.credo.aec_800g import CmisAec800gMemMap
+from .codes.credo.aec_800g import CredoAec800gCodes
+from .api.credo.aec_800g import CredoAec800gApi
+from .mem_maps.credo.aec_800g import CredoAec800gMemMap
 
 from .api.innolight.fr_800g import CmisFr800gApi
 from .api.hisense.aoc_2x100g import CmisAocSingleBankApi
@@ -85,7 +85,7 @@ class XcvrApiFactory(object):
         vendor_pn = self._get_vendor_part_num()
 
         if vendor_name == 'Credo' and vendor_pn in CREDO_800G_AEC_VENDOR_PN_LIST:
-            api = self._create_api(CmisAec800gCodes, CmisAec800gMemMap, CmisAec800gApi)
+            api = self._create_api(CredoAec800gCodes, CredoAec800gMemMap, CredoAec800gApi)
         elif ('INNOLIGHT' in vendor_name and vendor_pn in INL_800G_VENDOR_PN_LIST) or \
              ('EOPTOLINK' in vendor_name and vendor_pn in EOP_800G_VENDOR_PN_LIST):
             api = self._create_api(CmisCodes, CmisMemMap, CmisFr800gApi)
