+++ tests/sonic_xcvr/test_xcvr_api_factory.py
@@ -2,10 +2,10 @@ from unittest.mock import patch
 from mock import MagicMock
 import pytest
 
-from sonic_platform_base.sonic_xcvr.api.credo.aec_800g import CmisAec800gApi
-from sonic_platform_base.sonic_xcvr.mem_maps.credo.aec_800g import CmisAec800gMemMap
+from sonic_platform_base.sonic_xcvr.api.credo.aec_800g import CredoAec800gApi
+from sonic_platform_base.sonic_xcvr.mem_maps.credo.aec_800g import CredoAec800gMemMap
 from sonic_platform_base.sonic_xcvr.xcvr_eeprom import XcvrEeprom
-from sonic_platform_base.sonic_xcvr.codes.credo.aec_800g import CmisAec800gCodes
+from sonic_platform_base.sonic_xcvr.codes.credo.aec_800g import CredoAec800gCodes
 from sonic_platform_base.sonic_xcvr.api.innolight.fr_800g import CmisFr800gApi
 from sonic_platform_base.sonic_xcvr.api.hisense.aoc_2x100g import CmisAocSingleBankApi
 from sonic_platform_base.sonic_xcvr.api.amphenol.backplane import AmphBackplaneImpl
@@ -50,10 +50,10 @@ class TestXcvrApiFactory(object):
     @patch('sonic_platform_base.sonic_xcvr.xcvr_api_factory.XcvrApiFactory._get_vendor_part_num', MagicMock(return_value='CAC81X321M2MC1MS'))
     def test_create_xcvr_api(self):
         self.api.reader = self.mock_reader
-        CmisAec800gCodes = MagicMock()
-        CmisAec800gMemMap = MagicMock()
+        CredoAec800gCodes = MagicMock()
+        CredoAec800gMemMap = MagicMock()
         XcvrEeprom = MagicMock()
-        CmisAec800gApi = MagicMock()
+        CredoAec800gApi = MagicMock()
         self.api.create_xcvr_api()
 
     @patch('sonic_platform_base.sonic_xcvr.xcvr_api_factory.XcvrApiFactory._get_vendor_name', MagicMock(return_value='CISCO-INNOLIGHT'))
